+++ .github/workflows/deploy.yml
@@ -10,6 +10,7 @@ jobs:
       - uses: actions/setup-node@v3
         with:
           node-version: "18"
+          cache: npm
       - run: npm ci
       - run: npm run build
       - uses: JS-DevTools/npm-publish@v2
